pyproject.toml
[tool.poetry]
name = "tesktaskrlt"
version = "0.1.0"
description = ""
authors = ["GutesTod <jefflol003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.7.3"
aiogram = "^3.7.0"
python-dotenv = "^1.0.1"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
